+++ src/components/Heading/Heading.module.css
@@ -1,5 +1,5 @@
 .headingWrapper {
-  @apply min-w-[320px] w-[320px] h-[66px] md:min-w-[768px];
+  @apply min-w-[320px] max-w-4xl h-[66px] md:min-w-[768px];
   @apply flex justify-between items-center bg-secondary px-4 mx-auto;
 }
