+++ src/layout/RootLayout/RootLayout.jsx
@@ -2,8 +2,8 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { Toaster } from 'react-hot-toast';
 import { Outlet } from 'react-router-dom';
-import FooterBar from './FooterBar';
-import HeaderBar from './HeaderBar';
+import FooterBar from '../FooterBar';
+import HeaderBar from '../HeaderBar';
 
 const queryClient = new QueryClient();
 
@@ -11,7 +11,7 @@ function RootLayout() {
   return (
     <QueryClientProvider client={queryClient}>
       <HeaderBar />
-      <main className={`flex flex-1`}>
+      <main className={`flex`}>
         <div className="flex-1">
           <Outlet />
         </div>
