+++ src/routes.jsx
@@ -1,5 +1,5 @@
 import { createHashRouter } from 'react-router-dom';
-import RootLayout from './layout/RootLayout';
+import RootLayout from './layout/RootLayout/RootLayout';
 import Home from './pages/Home';
 import Post from './pages/Post/Post';
 import Writing from './pages/Writing';
